requirements.txt
crewai
crewai-tools
streamlit
python-dotenv
selenium
webdriver-manager